requirements.txt
fastapi
uvicorn
pytesseract
opencv-python
pillow
motor
python-multipart
jinja2
spacy 